Generate a question

Main function to generate a question in XML format. Combines various components such as type, text, and body to construct the complete XML question.

Usage

generate_question(
  first_question_number,
  copyright,
  license,
  author,
  correct_feedback,
  partially_correct_feedback,
  incorrect_feedback,
  adapt_images,
  width,
  height,
  type,
  question,
  image,
  image_alt,
  answer,
  ...
)

Arguments

first_question_number

An integer, the first number to compose the question names.

copyright

A string, copyright text to include in the question.

license

A string, license text to include in the question.

author

A string, author name.

correct_feedback

A string, feedback for correct answers.

partially_correct_feedback

A string, feedback for partially correct answers.

incorrect_feedback

A string, feedback for incorrect answers.

adapt_images

A boolean, whether to adapt images for size consistency.

width

An integer, width of the images.

height

An integer, height of the images.

type

A string, question type (if applicable).

question

A string, the main question text.

image

A string, optional image file to include in the question.

image_alt

A string, description of the image for accessibility.

answer

A string, correct answer for the question.

...

Additional strings, other possible answers.

Value

A string containing the question in XML format.