generate questiontext node

generate questiontext node

Usage

xml_questiontext(
  copyright,
  license,
  adapt_images,
  width,
  height,
  question,
  image,
  image_alt,
  type,
  author = "",
  fb_general = "",
  idnumber = ""
)

Arguments

copyright

A string, copyright text to be included in each question that is defined.

license

A string, license text to be included in each question that is defined.

adapt_images

A boolean, adapt the images so that they are a similar size.

width

A integer, width of each image.

height

A integer, height of each image.

question

A string, statement of the question.

image

A string, optional, image file to include in the question.

image_alt

A string, description of the image to include in the question.

type

A string, question type.

author

A string, author name to be included in each question that is defined.

fb_general

A string, general feedback to be included in each question that is defined.

idnumber

A string, idnumber to be included in each question that is defined.

Value

A string.